Refactor PlayerList into separate component
