+++ src/pages/GameMasterBoard.js
@@ -13,6 +13,7 @@ import {
   Divider,
 } from '@chakra-ui/react';
 import { Link, useParams } from 'react-router-dom';
+import PlayerList from '../components/gameMasterComponents/PlayerList';
 
 const GameMasterBoard = () => {
   const { roomID } = useParams();
@@ -89,25 +90,5 @@ const GameMasterBoard = () => {
     </Flex>
   );
 };
-
-const PlayerList = ({ title, players, taskCount, onPlayerDeath }) => (
-  <Box bg="gray.100" p={4} borderRadius="md" mb={3}>
-    <Heading size="md">{title}</Heading>
-    {players.map((player, index) => (
-      <Flex key={index} justify="space-between" mt={2}>
-        <Text fontSize="sm">{player}</Text>
-        {onPlayerDeath && (
-          <Button
-            colorScheme="red"
-            size="sm"
-            onClick={() => onPlayerDeath(player)}
-          >
-            Kill
-          </Button>
-        )}
-      </Flex>
-    ))}
-  </Box>
-);
 
 export default GameMasterBoard;
